requirements.txt
streamlit
folium
streamlit-folium
pandas
google-cloud-storage
streamlit-autorefresh
